streamlit/requirements.txt
streamlit==1.35.0
supabase==2.4.0
pandas==2.2.2
plotly==5.22.0
wordcloud==1.9.3
matplotlib==3.9.0
Pillow==10.3.0
python-dotenv==1.0.1